feat(absences): absence list is now the same as notes for better uniformization
