+++ lib/main_handler.dart
@@ -1,6 +1,6 @@
 import 'package:ent/screens/calendar_screen.dart';
 import 'package:ent/screens/home_screen.dart';
-import 'package:ent/screens/absence_screen.dart';
+import 'package:ent/screens/absences/absence_screen.dart';
 import 'package:ent/screens/notes/note_screen.dart';
 import 'package:ent/screens/settings_screen.dart';
 import 'package:flutter/material.dart';
