+++ lib/widgets/hamburger_menu.dart
@@ -5,7 +5,7 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import '../screens/calendar_screen.dart';
-import '../screens/absence_screen.dart';
+import '../screens/absences/absence_screen.dart';
 import '../screens/notes/note_screen.dart';
 
 class HamburgerMenu extends StatelessWidget {
